Call forward declared functions from templates in unordered tests.
Borland is having some issues with the existing tests, since they call
inline functions before they're defined. It might be right to, although
all the other compilers are fine with it. But the test isn't really what
the forward headers are intended for. Try to make the tests better
represent the intent, and possibly work on Borland.

git-svn-id: http://svn.boost.org/svn/boost/trunk@66428 b8fc166d-592f-0410-95f2-cb63ce0dd405